src/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic-settings==2.10.1
aiofiles ==24.1.0
langchain ==0.3.27
langchain-community
pypdf==6.0.0
PyPDF2==3.0.1
PyMuPDF==1.26.3
pdfplumber==0.11.7
python-docx==1.2.0
unstructured==0.18.13
docx2txt==0.8
langchain-text-splitters == 0.3.9
motor == 3.7.1
pydantic-mongo==3.1.0
openai==1.35.13
cohere==5.5.8
qdrant-client==1.15.1
torch == 2.8.0
transformers==4.56.0
sentence-transformers== 5.1.0
google-generativeai==0.8.5

# Monitoring and metrics
prometheus-client==0.21.1
starlette-exporter==0.23.0
fastapi-health==0.4.0